kbdbacklight: Listen for upower signal of changed brightness
Both the applet and OSD will want to use this to show updates when the
brightness changes.

There is no need to disable listening for this signal when the applet is
closed. We don't expect this signal to happen too frequently.

`Get` can still be used when showing the applet, as a fallback if there
are some backlights that don't generate this signal.