settings_daemon: Update brightness on name-owner change
Send events is settings daemon is started after the start of this
subscription, or if it is restarted.

To get updates when the daemon wasn't running at the start of the
stream, https://github.com/dbus2/zbus/pull/1389 is also needed.

Loss of the name owner should probably also disable any brightness
controls, but if that's done by a change of the brightness values and
not by a seperate property, we probably can't rely on the ordering of
events from different streams...